refactor(desktop): extract markdown spoiler renderer
Move the spoiler inline renderer into the markdown submodule without changing its implementation.
